Extract out useInboxDeepLinkListSync hook
